后端：新增common、data模块、新增data(mybatis)模块，连接mysql查询用户信息、整理各微服务架构与规范

+++ l-mix-boot/src/main/java/com/tornado4651/lmix/boot/domain/User.java
@@ -58,9 +58,9 @@ public class User implements Serializable {
     private String avatar;
 
     /**
-     * 用户状态（0：锁定；1：正常）
+     * 用户状态（0：锁定；1：正常; 3：暂时封禁）
      */
-    private Short status;
+    private Integer status;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
